+++ blocks/contact-form/contact-form.js
@@ -1,6 +1,4 @@
-import { loadContactForm } from '../../scripts/scripts.js';
-
 // eslint-disable-next-line no-unused-vars
-export default async function decorate(block) {
-  await loadContactForm(block);
+export default function decorate(block) {
+  // no-op
 }
