+++ scripts/delayed.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM, loadScript } from './lib-franklin.js';
+import { sampleRUM } from './lib-franklin.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
@@ -7,3 +7,19 @@ sampleRUM('cwv');
 // Adding timer for button click in the Hero Carousel
 
 // add more delayed functionality here
+
+const contactFormBlock = document.querySelector('.block.contact-form');
+if (contactFormBlock) {
+  const script = document.createElement('script');
+  script.setAttribute('type', 'text/javascript');
+  script.src = 'https://js.hsforms.net/forms/embed/v2.js';
+  script.addEventListener('load', () => {
+    // eslint-disable-next-line no-undef
+    hbspt.forms.create({
+      region: 'na1',
+      portalId: '14560092',
+      formId: 'af288827-9574-4a29-89eb-1b853eae1c20',
+    });
+  });
+  contactFormBlock.append(script);
+}
